# docs/swagger.yaml
basePath: /v1
definitions:
...
  backend-go_internal_core_models.Role:
    properties:
      RoleFeature:
        items:
          $ref: '#/definitions/backend-go_internal_core_models.RoleFeature'
        type: array
      active:
        type: boolean
      created_at:
        type: string
      deleted_at:
        type: string
      description:
        type: string
      id:
        type: string
      is_deleted:
        type: boolean
      name:
        type: string
      updated_at:
        type: string
    type: object
  backend-go_internal_core_models.RoleFeature:
    properties:
      activate:
        type: boolean
      create:
        type: boolean
      delete:
        type: boolean
      id_feature:
        type: string
      id_role:
        type: string
      view:
        type: boolean
    type: object
  backend-go_internal_core_models.User:
    properties:
      Role:
        $ref: '#/definitions/backend-go_internal_core_models.Role'
      active:
        type: boolean
      avatar:
        type: string
      cognito_id:
        type: string
      created_at:
        type: string
      deleted_at:
        type: string
      document:
        type: string
      email:
        type: string
      id:
        type: string
      id_auth:
        type: string
      id_role:
        type: string
      is_deleted:
        type: boolean
      name:
        type: string
      phone:
        type: string
      updated_at:
        type: string
    type: object
...
  internal_app_user.CreateUserDTO:
    properties:
      active:
        type: boolean
      email:
        type: string
      id_role:
        type: string
      name:
        type: string
      password:
        type: string
    required:
    - email
    - id_role
    - name
    - password
    type: object
  internal_app_user.UpdateStatusRequest:
    properties:
      active:
        type: boolean
    type: object
  internal_app_user.UpdateUserDTO:
    properties:
      active:
        type: boolean
      email:
        type: string
      id_role:
        type: string
      name:
        type: string
      password:
        type: string
    type: object
  internal_app_user.UserListAllResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/backend-go_internal_core_models.User'
        type: array
      total:
        type: integer
    type: object
  internal_app_user.UserListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/backend-go_internal_core_models.User'
        type: array
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
    type: object
host: localhost:8888
info:
  contact:
    email: suporte@swagger.io
    name: Suporte API
    url: http://www.swagger.io/support
  description: API modular em Go com Gin e Swagger.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Backend Go API
  version: "1.0"
paths:
...
  /user:
    get:
      description: Retorna usuários com paginação e filtros.
      parameters:
      - description: 'Número da página (padrão: 1)'
        in: query
        name: page
        type: integer
      - description: 'Itens por página (padrão: 25)'
        in: query
        name: size
        type: integer
      - description: Termo de busca
        in: query
        name: searchWord
        type: string
      - description: 'Campos para busca (ex: name,email,Role.name)'
        in: query
        name: searchFields
        type: string
      - description: Campo para ordenação
        in: query
        name: orderBy
        type: string
      - description: Direção da ordenação (asc/desc)
        in: query
        name: orderDirection
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_app_user.UserListResponse'
      security:
      - Bearer: []
      summary: Listar Usuários
      tags:
      - User
    post:
      consumes:
      - application/json
      description: Cria um novo usuário com os dados fornecidos.
      parameters:
      - description: Dados do usuário
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_app_user.CreateUserDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/backend-go_internal_core_models.User'
        "400":
          description: Dados inválidos
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Erro interno
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Criar Usuário
      tags:
      - User
  /user/{id}:
    delete:
      description: Remove um usuário logicamente baseado no ID.
      parameters:
      - description: ID do Usuário
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 'message: usuário excluído com sucesso'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Erro interno
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Excluir Usuário
      tags:
      - User
    get:
      description: Retorna os detalhes de um usuário específico.
      parameters:
      - description: ID do Usuário
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/backend-go_internal_core_models.User'
        "404":
          description: Usuário não encontrado
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Obter Usuário
      tags:
      - User
    put:
      consumes:
      - application/json
      description: Atualiza os dados de um usuário baseado no ID.
      parameters:
      - description: ID do Usuário
        in: path
        name: id
        required: true
        type: string
      - description: Campos a serem atualizados
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_app_user.UpdateUserDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/backend-go_internal_core_models.User'
        "400":
          description: Dados inválidos
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Erro interno
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Atualizar Usuário
      tags:
      - User
  /user/{id}/status:
    patch:
      consumes:
      - application/json
      description: Define se um usuário está ativo ou inativo.
      parameters:
      - description: ID do Usuário
        in: path
        name: id
        required: true
        type: string
      - description: Novo status
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_app_user.UpdateStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 'message: status atualizado com sucesso'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Erro interno
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Alterar Status do Usuário
      tags:
      - User
  /user/all:
    get:
      description: Retorna todos os usuários (usado para selects/lookups).
      parameters:
      - description: Termo de busca
        in: query
        name: searchWord
        type: string
      - description: Campos para busca
        in: query
        name: searchFields
        type: string
      - description: Campo para ordenação
        in: query
        name: orderBy
        type: string
      - description: Direção da ordenação
        in: query
        name: orderDirection
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_app_user.UserListAllResponse'
      security:
      - Bearer: []
      summary: Listar Todos os Usuários
      tags:
      - User
  /user/export/pdf:
    get:
      description: Exporta a lista filtrada de usuários para um arquivo PDF.
      parameters:
      - description: Termo de busca
        in: query
        name: searchWord
        type: string
      - description: Campos para busca
        in: query
        name: searchFields
        type: string
      - description: Campo para ordenação
        in: query
        name: orderBy
        type: string
      - description: Direção da ordenação
        in: query
        name: orderDirection
        type: string
      produces:
      - application/pdf
      responses:
        "200":
          description: PDF file
          schema:
            type: file
        "400":
          description: Dados inválidos
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Erro interno
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Exportar Usuários para PDF
      tags:
      - User
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"